feat: add deep copy constructor in Forest class
Deep copy will create a safe instance copy to another Forest object, that will
be used to create a buffer for the threads save their results.

This constructor receive the original object and create a new instance
with the original values, creating a safe copy that will no interfere in the
original instance.