feat: adds SimulationTask for Parallel Processing of Fire Propagation Calculation
Implements the SimulationTask class, which adopts the Callable<Forest> interface
to encapsulate the fire propagation logic within a thread.
This implementation is to enabling parallel processing of the forest matrix, where
every thread will work only in one portion of the grid.

The class solves the concurrency problem with the following strategy:

1. **currentForest:** Used only for reading data, so the thread will not have
the problem of reading a cell that was modified by another thread.

2. **bufferForest:** Used exclusively for modifying the cell state.

The `calculateIgnitionProbability()` returns an probability value in range
of 0% to 100%. To get a probability more realistic and avoid physics complex
equations, the method uses this logic:

- **Base probability of ignition (baseProb):** increase with the number of
neighbors burnig and can decrease with the moisture level of the target cell.

- **Wind factor (windProb):** Is considered if one of the burning neighbors is
in the way of the Wind Direction. The speed of the wind will increase or decrease
the percentage of influency that the wind will have in the final probability.

- **Relief Factor (reliefProb):** will increase in 5% for each burning neighbor
that has a relief lower than the target cell's relief.

The method sums the 3 factors and normalizes to 100% or 0% if the total
value is greater or less than these limits.

With this return, the SimulationTask will use the ´Monte Carlo Method´ to
randomize the probability. So, if the random double (0.0 to 1.0) is lower
than the probability of ignition, the cell will start to burn.