feat: adds calculateNextGeneration() with threads
This method is used to implements the parallel mode to scan the matrix
and apply new cell states using the automatum celullar automata logic.

To solve this problem, the method use this steps:

1. **SimulationTask: ** creates and store instances of SimulationTask in a
List. Each task receives the range of index to work and two Forests:
buffer to write and original to read.

2. **Executor: ** With executor, the method invokeAll uses the threads to
get tasks from the list and solve.

3. **BufferForest instance:** as final step, the currentForest will
receive the instance of BufferForest to store the new grid for another cycle.