fix: adds verification to update cell that is already burning
The logic was not allowing the burning cells turn into ash because the timer
at model was not being updated.

To solve this, the condition structure will check the cell. If its already
burning, then will use the method at model to updateBurningState before
continue to another cell.

The updateBurningState will decrease the timer until 0. If reachs 0, then
the cell will turn into ash.